perf(v4flash-kernels): f16 V + f16 scores + q2k by-expert + qb WMMA — +13% e2e prefill at 64K
Four landed rollouts, cumulative +13% on a single-chunk 64K prefill
trace (1666 ms/chunk span → 1450 ms/chunk).

1. f16 V end-to-end. raw_kv + comp_kv now u16; attention readers cast at
   use. smwsum_ldsv 14.3 → 13.1 ms (-9%) in production. Halves the V
   cache memory. Same RTNE rounding the WMMA kernels already did just
   before the WMMA op — no precision regression. max_abs vs ds4 = 2.86e-6
   through SWA output.

2. f16 scores end-to-end. New attention_mixed_softmax_wsum_batched_htiled
   _wmma_ldsv_f16s smwsum kernel combines LDS-V staging with f16 score
   reads. Isolated smwsum 13.05 → 10.09 ms (-23%) at depth 32K, B=256.
   E2E +2.9% at 64K. Score scratch (bd.attn_scores) halved to
   N_HEAD * ATTN_MIXED_MAX_KEYS/2 f32 elements (≡ same bytes as f16) —
   saves ~290 MB at B_MAX=512.

3. q2k_down by-expert. Inverts the (B, expert) iteration of
   q2_k_matvec_par_batched_BxN: each expert's row-tile read ONCE per WG
   instead of B*n_used times. Reuses the iq2 group_count/expert_members/
   work_items arrays — no extra pre-pass. Writes per-(b, slot) partials
   to bi.q2k_partials (no atomic — deterministic, run-to-run bit-identical),
   then a tiny q2_k_reduce_partials sums across n_used per (b, row).
   Was 94% iGPU DRAM-BW-bound; now amortized. E2E +14-23% prefill
   across depths at B_MAX=512 (crossover ≈ B=64).
   Q2K_VARIANT=bxn rolls back.

4. qb WMMA default ON. q8_0_gemm_wmma_i8 (int8-WMMA) replaces dp4a
   matvec_batched for the qb up-projection. 2.0-2.5× isolated win at
   M=Q_FLAT=32768 across B={64,256,512}. Was opt-in via QB_WMMA=1;
   e2e was neutral when iGPU MoE was binding (pre-q2k-by-expert).
   Post-q2k-by-expert, dGPU became the binding side and qb WMMA recovers
   +2% e2e at 64K. QB_WMMA=0 rolls back.

Removes the ATTN_SCORES_F32 / ATTN_SMWSUM_DRAM_V env rollbacks (the
half-size scratch is incompatible with the f32-scores chain — clean cut
since the f16s chain is correctness-verified and bench-verified).

Bench harness additions: bench_q2k_isolated (new, A/B + correctness vs
BxN), bench_qb_wmma_isolated takes BENCH_M/BENCH_K for arbitrary GEMM
shape sweep, bench_iq2_isolated ATT_SKIP_FILL=1 to make iq2 the first
dispatched kernel (rocprofv3 --att-consecutive-kernels=1 catches it
without needing a regex filter, which hangs ATT on gfx1151 in 7.2.3).

+++ crates/v4flash-kernels/src/comp_kv_append.rs
@@ -24,10 +24,12 @@ impl CompKvAppend {
         Ok(Self { module })
     }
 
+    /// `comp_kv` is `[n_comp_max * head_dim]` f16 (held as `u16`).
+    /// `row` is `[head_dim]` f32 (cast to f16 at store).
     pub fn launch(
         &self,
         stream: &Stream,
-        comp_kv: &mut DeviceBuffer<f32>,
+        comp_kv: &mut DeviceBuffer<u16>,
         row: &DeviceBuffer<f32>,
         n_comp: u32,
         head_dim: u32,
